+++ .eslintrc.js
@@ -1,0 +1,15 @@
+module.exports = {
+  root: true,
+  env: {
+    node: true
+  },
+  ignorePatterns: ['node_modules/*', 'dist/*'],
+  extends: ['plugin:vue/essential', 'eslint:recommended', '@vue/prettier'],
+  parserOptions: {
+    parser: 'babel-eslint'
+  },
+  rules: {
+    'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
+    'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off'
+  }
+}
