+++ LICENSE
@@ -1,6 +1,4 @@
-MIT License
-
-Copyright (c) 2018 Freerk van Zeijl
+Copyright (c) 2018 Jan Grzegorowski
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
