+++ examples/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <meta http-equiv="x-ua-compatible" content="ie=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>VueBadgerAccordion</title>
+    <script src="https://cdn.jsdelivr.net/npm/vue"></script>
+    <script src="../dist/vue-badger-accordion.js" defer></script>
+  </head>
+  <body>
+    <div id="app"></div>
+  </body>
+</html>
