--- index.js
@@ -1,5 +1,0 @@
-import ImportBadgerAccordion from './src/BadgerAccordion'
-import ImportBadgerAccordionItem from './src/BadgerAccordionItem'
-
-export const BadgerAccordion = ImportBadgerAccordion;
-export const BadgerAccordionItem = ImportBadgerAccordionItem;
