+++ package.json
@@ -1,19 +1,27 @@
 {
   "name": "vue-badger-accordion",
-  "version": "1.1.5",
-  "description": "Badger-Accordion Component for Vue 2.0",
-  "main": "index.js",
+  "version": "2.0",
+  "description": "Badger-Accordion Component for Vue",
+  "main": "src/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "clean": "rimraf dist",
+    "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
+    "lintfix": "eslint --fix --ext .js,.vue --ignore-path .gitignore .",
+    "build": "yarn clean && yarn lint && yarn rollup -c"
   },
+  "engines": {
+    "node": ">= 8.x"
+  },
+  "private": false,
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/vanderb/vue-badger-accordion.git"
+    "url": "https://github.com/vanderb/vue-badger-accordion.git"
   },
   "keywords": [
     "vue",
     "accordion",
-    "collapse"
+    "collapse",
+    "badger"
   ],
   "author": "Freerk van Zeijl",
   "license": "MIT",
@@ -22,7 +30,39 @@
   },
   "homepage": "https://github.com/vanderb/vue-badger-accordion#readme",
   "dependencies": {
-    "badger-accordion": "^1.2.2",
-    "vue": "^2.5.17"
+    "badger-accordion": "^1.2.4",
+    "core-js": "^3.6.5",
+    "vue": "^2.6.11"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.12.0",
+    "@babel/parser": "^7.12.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/preset-env": "^7.12.0",
+    "@rollup/plugin-babel": "^5.2.1",
+    "@rollup/plugin-commonjs": "^15.1.0",
+    "@rollup/plugin-node-resolve": "^9.0.0",
+    "@vue/compiler-sfc": "^3.0.1",
+    "@vue/eslint-config-prettier": "^6.0.0",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.0.0-beta",
+    "cross-env": "^7.0.2",
+    "css-loader": "^5.0.0",
+    "eslint": "^6.7.2",
+    "eslint-plugin-prettier": "^3.1.3",
+    "eslint-plugin-vue": "^6.2.2",
+    "html-webpack-plugin": "^4.5.0",
+    "node-sass": "^4.14.1",
+    "prettier": "^1.19.1",
+    "rimraf": "^3.0.2",
+    "rollup": "^2.31.0",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-css-only": "^2.1.0",
+    "rollup-plugin-node-resolve": "^5.2.0",
+    "rollup-plugin-prettier": "^2.1.0",
+    "rollup-plugin-scss": "^2.6.1",
+    "rollup-plugin-vue": "^6.0.0-beta.10",
+    "sass-loader": "^10.0.3",
+    "webpack": "^5.1.2"
   }
 }
