+++ types/vue-badger-accordion.d.ts
@@ -1,6 +1,6 @@
-declare module 'vue-badger-accordion' {
-    import Vue from 'vue';
-    class BadgerAccordion extends Vue { }
-    class BadgerAccordionItem extends Vue { }
-    export { BadgerAccordion, BadgerAccordionItem }
+declare module "vue-badger-accordion" {
+  import Vue from 'vue';
+  class BadgerAccordion extends Vue {}
+  class BadgerAccordionItem extends Vue {}
+  export { BadgerAccordion, BadgerAccordionItem };
 }
